requirements.txt
streamlit
openai
pandas
numpy
matplotlib
python-dotenv
